move redirectFirst from lib/view/pagination to lib/tournament
Per ornicar's review on #19512, redirectFirst is not related to
views or pagination. Move it to its own lib/tournament module.

Entire-Checkpoint: ff7d63c03877
